[tests] Improved test infrastructure - Make clear what is host code and what is cluster code
