+++ tests/generic/snitchCluster/simpleOffload/include/test_cluster.h
@@ -1,0 +1,16 @@
+// Copyright 2024 ETH Zurich and University of Bologna.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+//
+// Philip Wiese <wiesep@iis.ee.ethz.ch>
+
+#ifndef _TEST_CLUSTER_INCLUDE_GUARD_
+#define _TEST_CLUSTER_INCLUDE_GUARD_
+
+#include <stdint.h>
+
+void clusterInterruptHandler();
+
+int32_t testReturn(void *args);
+
+#endif //_TEST_CLUSTER_INCLUDE_GUARD_
